Создал отдельные компоненты конструктора
