+++ frontend/src/common/components/AppDrag.vue
@@ -5,7 +5,7 @@
         @dragover.prevent
         @dragenter.prevent
     >
-
+        <slot />
     </div>
 </template>
   
@@ -23,6 +23,7 @@
     })
 
     const onDragStart = ({ dataTransfer }) => {
+        console.log(dataTransfer)
         const data = JSON.stringify(props.dataTransfer);
         dataTransfer.setData(DATA_TRANSFER_PAYLOAD, data);
     }
