+++ frontend/src/common/constants/index.js
@@ -1,0 +1,11 @@
+export const DATA_TRANSFER_PAYLOAD = "payload";
+
+export const CounterTypes = {
+    DecrementType: "minus",
+    IncrementType: "plus" 
+};
+
+export const CounterValue = {
+    DecrementValue: "Меньше",
+    IncrementValue: "Больше",
+}
