--- frontend/src/common/contants/index.js
@@ -1,1 +1,0 @@
-export const DATA_TRANSFER_PAYLOAD = "payload";
