+++ frontend/src/modules/constructor/Pizza.vue
@@ -1,0 +1,343 @@
+<template>
+    <app-drop @drop="$emit('drop', $event)">
+      <div class="content__pizza">
+          <label class="input">
+              <span class="visually-hidden">Название пиццы</span>
+              <input type="text" name="pizza_name" placeholder="Введите название пиццы">
+          </label>
+      
+          <div class="content__constructor">
+              <div class="pizza pizza--foundation--small-tomato">
+                  <div class="pizza__wrapper">
+                      <div class="pizza__filling pizza__filling--ananas"></div>
+                      <div class="pizza__filling pizza__filling--bacon"></div>
+                      <div class="pizza__filling pizza__filling--cheddar"></div>                    
+                  </div>
+              </div>
+          </div>
+      
+          <div class="content__result">
+              <p>Итого: 0 ₽</p>
+              <button type="button" class="button" disabled>Готовьте!</button>
+          </div>
+      </div>
+    </app-drop>  
+</template>
+
+<script setup>
+  import AppDrop from '@/common/components/AppDrop.vue';
+  
+  defineEmits(['drop'])
+</script>
+
+<style scoped lang="scss">
+    @import '@/assets/scss/ds-system/ds.scss';
+    @import '@/assets/scss/mixins/mixins.scss';
+    @import '@/assets/scss/visually-hidden.scss';
+    .content__pizza {
+        width: 373px;
+        margin-top: 15px;
+        margin-bottom: 15px;
+    }
+    .content__constructor {
+        width: 315px;
+        margin-top: 25px;
+        margin-right: auto;
+        margin-left: auto;
+    }
+
+    .content__wrapper {
+        display: flex;
+        align-items: flex-start;
+        flex-wrap: wrap;
+        width: 920px;
+        margin: 0 auto;
+        padding-right: 2.12%;
+        padding-bottom: 30px;
+        padding-left: 2.12%;
+    }
+
+    .content__result {
+        display: flex;
+        align-items: center;
+        justify-content: center;
+        margin-top: 25px;
+        p {
+            @include b-s24-h28;
+            margin: 0;
+        }
+        button {
+            margin-left: 12px;
+            padding: 16px 45px;
+        }
+    }
+    .input {
+        display: block;
+        span {
+            @include r-s14-h16;
+            display: block;
+            margin-bottom: 4px;
+        }
+        input {
+            @include r-s16-h19;
+            display: block;
+            box-sizing: border-box;
+            width: 100%;
+            margin: 0;
+            padding: 8px 16px;
+            transition: 0.3s;
+            color: $black;
+            border: 1px solid $purple-400;
+            border-radius: 8px;
+            outline: none;
+            background-color: $white;
+            font-family: inherit;
+            &:focus {
+                border-color: $green-500;
+            }
+        }
+    &:hover {
+        input {
+            border-color: $black;
+        }
+    }
+    &--big-label {
+        display: flex;
+        align-items: center;
+        span {
+            @include b-s16-h19;
+            margin-right: 16px;
+            white-space: nowrap;
+        }
+    }
+  }
+
+  .button {
+    $bl: &;
+    @include b-s18-h21;
+    font-family: inherit;
+    display: block;
+    box-sizing: border-box;
+    margin: 0;
+    padding: 0;
+    cursor: pointer;
+    transition: 0.3s;
+    text-align: center;
+    color: $white;
+    border: none;
+    border-radius: 8px;
+    outline: none;
+    box-shadow: $shadow-medium;
+    background-color: $green-500;
+    &:hover:not(:active):not(:disabled) {
+      background-color: $green-400;
+    }
+    &:active:not(:disabled) {
+      background-color: $green-600;
+    }
+    &:focus:not(:disabled) {
+      opacity: 0.5;
+    }
+    &:disabled {
+      background-color: $green-300;
+      color: rgba($white, 0.2);
+      cursor: default;
+    }
+    &--border {
+      background-color: transparent;
+      border: 1px solid $green-500;
+      color: $black;
+      box-shadow: none;
+      &:hover:not(:active):not(:disabled) {
+        color: $green-500;
+        border-color: $green-500;
+        background-color: transparent;
+      }
+      &:active:not(:disabled) {
+        color: $green-600;
+        border-color: $green-600;
+        background-color: transparent;
+      }
+      &:disabled {
+        opacity: 0.5;
+      }
+    }
+    &--transparent {
+      @include b-s14-h16;
+      background-color: transparent;
+      box-shadow: none;
+      color: $black;
+      &:hover:not(:active):not(:disabled) {
+        color: $red-800;
+        background-color: transparent;
+      }
+      &:active:not(:disabled) {
+        color: $red-900;
+        background-color: transparent;
+      }
+      &:disabled {
+        opacity: 0.25;
+      }
+    }
+    &--arrow {
+      &::before {
+        content: "";
+        background-image: url("@/assets/img/button-arrow.svg");
+        background-position: center;
+        background-repeat: no-repeat;
+        margin-right: 16px;
+        width: 18px;
+        height: 18px;
+        display: inline-block;
+        vertical-align: middle;
+        transform: translateY(-1px);
+      }
+    }
+    &--white {
+      background-color: $white;
+      color: $green-500;
+    }
+  }
+
+  .pizza {
+    position: relative;
+    display: block;
+    box-sizing: border-box;
+    width: 100%;
+    background-repeat: no-repeat;
+    background-position: center;
+    background-size: 100%;
+    &--foundation--big-creamy {
+      background-image: url("@/assets/img/foundation/big-creamy.svg");
+    }
+    &--foundation--big-tomato {
+      background-image: url("@/assets/img/foundation/big-tomato.svg");
+    }
+    &--foundation--small-creamy {
+      background-image: url("@/assets/img/foundation/small-creamy.svg");
+    }
+    &--foundation--small-tomato {
+      background-image: url("@/assets/img/foundation/small-tomato.svg");
+    }
+  }
+  .pizza__wrapper {
+    width: 100%;
+    padding-bottom: 100%;
+  }
+  .pizza__filling {
+    $bl: &;
+    position: absolute;
+    top: 0;
+    left: 0;
+    display: block;
+    width: 100%;
+    height: 100%;
+    background-repeat: no-repeat;
+    background-position: center;
+    background-size: 100%;
+    &::before,
+    &::after {
+      display: none;
+      position: absolute;
+      top: 0;
+      left: 0;
+      width: 100%;
+      height: 100%;
+      content: "";
+      background-image: inherit;
+    }
+    &--second {
+      &::before {
+        display: block;
+        transform: rotate(45deg);
+      }
+    }
+    &--third {
+      &::before {
+        display: block;
+        transform: rotate(45deg);
+      }
+      &::after {
+        display: block;
+        transform: rotate(-45deg);
+      }
+    }
+    &--ananas,
+    &--ananas.pizza__filling--second::before,
+    &--ananas.pizza__filling--third::after {
+      background-image: url("@/assets/img/filling-big/ananas.svg");
+    }
+    &--bacon,
+    &--bacon.pizza__filling--second::before,
+    &--bacon.pizza__filling--third::after {
+      background-image: url("@/assets/img/filling-big/bacon.svg");
+    }
+    &--blue_cheese,
+    &--blue.pizza__filling--second::before,
+    &--blue.pizza__filling--third::after {
+      background-image: url("@/assets/img/filling-big/blue_cheese.svg");
+    }
+    &--cheddar,
+    &--cheddar.pizza__filling--second::before,
+    &--cheddar.pizza__filling--third::after {
+      background-image: url("@/assets/img/filling-big/cheddar.svg");
+    }
+    &--chile,
+    &--chile.pizza__filling--second::before,
+    &--chile.pizza__filling--third::after {
+      background-image: url("@/assets/img/filling-big/chile.svg");
+    }
+    &--ham,
+    &--ham.pizza__filling--second::before,
+    &--ham.pizza__filling--third::after {
+      background-image: url("@/assets/img/filling-big/ham.svg");
+    }
+    &--jalapeno,
+    &--jalapeno.pizza__filling--second::before,
+    &--jalapeno.pizza__filling--third::after {
+      background-image: url("@/assets/img/filling-big/jalapeno.svg");
+    }
+    &--mozzarella,
+    &--mozzarella.pizza__filling--second::before,
+    &--mozzarella.pizza__filling--third::after {
+      background-image: url("@/assets/img/filling-big/mozzarella.svg");
+    }
+    &--mushrooms,
+    &--mushrooms.pizza__filling--second::before,
+    &--mushrooms.pizza__filling--third::after {
+      background-image: url("@/assets/img/filling-big/mushrooms.svg");
+    }
+    &--olives,
+    &--olives.pizza__filling--second::before,
+    &--olives.pizza__filling--third::after {
+      background-image: url("@/assets/img/filling-big/olives.svg");
+    }
+    &--onion,
+    &--onion.pizza__filling--second::before,
+    &--onion.pizza__filling--third::after {
+      background-image: url("@/assets/img/filling-big/onion.svg");
+    }
+    &--parmesan,
+    &--parmesan.pizza__filling--second::before,
+    &--parmesan.pizza__filling--third::after {
+      background-image: url("@/assets/img/filling-big/parmesan.svg");
+    }
+    &--salami,
+    &---salami.pizza__filling--second::before,
+    &---salami.pizza__filling--third::after {
+      background-image: url("@/assets/img/filling-big/salami.svg");
+    }
+    &--salmon,
+    &--salmon.pizza__filling--second::before,
+    &--salmon.pizza__filling--third::after {
+      background-image: url("@/assets/img/filling-big/salmon.svg");
+    }
+    &--tomatoes,
+    &--tomatoes.pizza__filling--second::before,
+    &--tomatoes.pizza__filling--third::after {
+      background-image: url("@/assets/img/filling-big/tomatoes.svg");
+    }
+  }
+
+
+</style>
