+++ frontend/src/views/HomeView.vue
@@ -5,106 +5,25 @@
           <div class="content__wrapper">
             <h1 class="title title--big">Конструктор пиццы</h1>
     
-            <div class="content__dough">
+            <dough />
     
-              <div class="sheet">
-                <h2 class="title title--small sheet__title">Выберите тесто</h2>
-    
-                <div class="sheet__content dough" >
-                  <label v-for="doughItem in dough" :class="`dough__input dough__input--${doughItem.value}`" :key="doughItem.id">
-                    <input type="radio" name="dought" :value="doughItem.value" class="visually-hidden">
-                    <img :src="getImage(doughItem.image)" :alt="doughItem.name" />              
-                    <b>{{ doughItem.name }}</b>
-                    <span>{{ doughItem.description }}</span>
-                  </label>
-                </div>
-    
-              </div>
-    
-            </div>
-    
-            <div class="content__diameter">
-              <div class="sheet">
-                <h2 class="title title--small sheet__title">Выберите размер</h2>
-    
-                <div class="sheet__content diameter">
-                  <label v-for="size in sizes" :class="`diameter__input diameter_input--${size.value}`" :key="size.id">
-                    <input type="radio" name="diameter" :value="size.value" class="visually-hidden">
-                    <span>{{ size.name }}</span>
-                  </label>
-                </div>
-              </div>
-            </div>
-    
+            <diameter />
+
             <div class="content__ingredients">
               <div class="sheet">
                 <h2 class="title title--small sheet__title">Выберите ингредиенты</h2>
     
                 <div class="sheet__content ingredients">
     
-                  <div class="ingredients__sauce">
-                    <p>Основной соус:</p>
+                  <sauce />
     
-                    <label v-for="sauce in sauces" :key="sauce.id" class="radio ingredients__input">
-                      <input type="radio" name="sauce" :value="sauce.value">
-                      <span>{{ sauce.name }}</span>
-                    </label>
-                  </div>
-    
-                  <div class="ingredients__filling">
-                    <p>Начинка:</p>
-    
-                    <ul class="ingredients__list">
-                      <li class="ingredients__item" v-for="ingridient in ingridients" :key="ingridient.id">
-                        <div :class="`filling filling--${ingridient.value}`">
-                          <img
-                            :src="getImage(ingridient.image)"
-                            :alt="ingridient.name"
-                            width="30"
-                            height="30"
-                          />
-                          {{ ingridient.name }}
-                        </div>
-    
-                        <div class="counter counter--orange ingredients__counter">
-                          <button type="button" class="counter__button counter__button--minus">
-                            <span class="visually-hidden">Меньше</span>
-                          </button>
-                          <input type="text" name="counter" class="counter__input" value="0">
-                          <button type="button" class="counter__button counter__button--plus">
-                            <span class="visually-hidden">Больше</span>
-                          </button>
-                        </div>
-                      </li>
-                    </ul>
-    
-                  </div>
+                  <ingredient-list />
     
                 </div>
               </div>
             </div>
     
-            <div class="content__pizza">
-              <label class="input">
-                <span class="visually-hidden">Название пиццы</span>
-                <input type="text" name="pizza_name" placeholder="Введите название пиццы">
-              </label>
-    
-              <div class="content__constructor">
-                <div class="pizza pizza--foundation--big-tomato">
-                  <div class="pizza__wrapper">
-                    <div class="pizza__filling pizza__filling--ananas"></div>
-                    <div class="pizza__filling pizza__filling--bacon"></div>
-                    <div class="pizza__filling pizza__filling--cheddar"></div>                    
-                  </div>
-                </div>
-              </div>
-    
-              <div class="content__result">
-                <p>Итого: 0 ₽</p>
-                <button type="button" class="button" disabled>Готовьте!</button>
-              </div>
-            </div>
+            <pizza />
     
           </div>
     
@@ -113,23 +32,11 @@
 </template>
 
 <script setup>
-import doughRow from '../mocks/dough.json';
-import ingridientsRow from '../mocks/ingredients.json';
-import saucesRow from '../mocks/sauces.json';
-import sizesRow from '../mocks/sizes.json';
-
-import { 
-  normalizeSize, 
-  normalizeIngridient, 
-  normalizeSauce, 
-  normalizeDough,
-  getImage
-} from '../common/helpers/normalize'
-
-const dough = doughRow.map(normalizeDough);
-const sizes = sizesRow.map(normalizeSize);
-const sauces = saucesRow.map(normalizeSauce);
-const ingridients = ingridientsRow.map(normalizeIngridient);
+import Dough from "@/modules/constructor/Dough.vue";
+import Diameter from "@/modules/constructor/Diameter.vue";
+import Sauce from "@/modules/constructor/Sauce.vue";
+import IngredientList from "@/modules/constructor/IngredientList.vue";
+import Pizza from '@/modules/constructor/Pizza.vue';
 
 </script>
 
@@ -151,48 +58,14 @@ const ingridients = ingridientsRow.map(normalizeIngridient);
     padding-bottom: 30px;
     padding-left: 2.12%;
   }
-  .content__dough {
-    width: 527px;
-    margin-top: 15px;
-    margin-right: auto;
-    margin-bottom: 15px;
-  }
-  .content__diameter {
-    width: 373px;
-    margin-top: 15px;
-    margin-bottom: 15px;
-  }
+
   .content__ingredients {
     width: 527px;
     margin-top: 15px;
     margin-right: auto;
     margin-bottom: 15px;
   }
-  .content__pizza {
-    width: 373px;
-    margin-top: 15px;
-    margin-bottom: 15px;
-  }
-  .content__constructor {
-    width: 315px;
-    margin-top: 25px;
-    margin-right: auto;
-    margin-left: auto;
-  }
-  .content__result {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    margin-top: 25px;
-    p {
-      @include b-s24-h28;
-      margin: 0;
-    }
-    button {
-      margin-left: 12px;
-      padding: 16px 45px;
-    }
-  }
+  
   .sheet {
     padding-top: 15px;
     border-radius: 8px;
@@ -213,31 +86,7 @@ const ingridients = ingridientsRow.map(normalizeIngridient);
     padding-left: 18px;
     border-top: 1px solid rgba($green-500, 0.1);
   }
-  .ingredients__sauce {
-    display: flex;
-    align-items: center;
-    flex-wrap: wrap;
-    width: 100%;
-    margin-bottom: 14px;
-    p {
-      @include r-s16-h19;
-      margin-top: 0;
-      margin-right: 16px;
-      margin-bottom: 10px;
-    }
-  }
-  .ingredients__input {
-    margin-right: 24px;
-    margin-bottom: 10px;
-  }
-  .ingredients__filling {
-    width: 100%;
-    p {
-      @include r-s16-h19;
-      margin-top: 0;
-      margin-bottom: 16px;
-    }
-  }
+ 
   .title {
     box-sizing: border-box;
     width: 100%;
@@ -250,540 +99,5 @@ const ingridients = ingridientsRow.map(normalizeIngridient);
       @include b-s18-h21;
     }
   }
-  .ingredients__list {
-    @include clear-list;
-    display: flex;
-    align-items: flex-start;
-    flex-wrap: wrap;
-  }
-  .ingredients__item {
-    width: 100px;
-    min-height: 40px;
-    margin-right: 17px;
-    margin-bottom: 35px;
-  }
-  .ingredients__counter {
-    width: 54px;
-    margin-top: 10px;
-    margin-left: 36px;
-  }
-  .radio {
-    cursor: pointer;
-    span {
-      @include r-s16-h19;
-      position: relative;
-      padding-left: 28px;
-      &:before {
-        @include p_center-v;
-        display: block;
-        box-sizing: border-box;
-        width: 20px;
-        height: 20px;
-        content: "";
-        transition: 0.3s;
-        border: 1px solid $purple-400;
-        border-radius: 50%;
-        background-color: $white;
-      }
-    }
-    &:hover {
-      input:not(:checked):not(:disabled) + span {
-        &:before {
-          border-color: $purple-800;
-        }
-      }
-    }
-    input {
-      display: none;
-      &:checked + span {
-        &:before {
-          border: 6px solid $green-500;
-        }
-      }
-      &:disabled {
-        & + span {
-          &:before {
-            border-color: $purple-400;
-            background-color: $silver-200;
-          }
-        }
-        &:checked + span {
-          &:before {
-            border: 6px solid $purple-400;
-          }
-        }
-      }
-    }
-  }
-  .counter {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-  }
-  .counter__button {
-    $el: &;
-    $size_icon: 50%;
-    position: relative;
-    display: block;
-    width: 16px;
-    height: 16px;
-    margin: 0;
-    padding: 0;
-    cursor: pointer;
-    transition: 0.3s;
-    border: none;
-    border-radius: 50%;
-    outline: none;
-    &--minus {
-      background-color: $purple-100;
-      &::before {
-        @include p_center-all;
-        width: $size_icon;
-        height: 2px;
-        content: "";
-        border-radius: 2px;
-        background-color: $black;
-      }
-      &:hover:not(:active):not(:disabled) {
-        background-color: $purple-200;
-      }
-      &:active:not(:disabled) {
-        background-color: $purple-300;
-      }
-      &:focus:not(:disabled) {
-        box-shadow: $shadow-regular;
-      }
-      &:disabled {
-        cursor: default;
-        &::before {
-          opacity: 0.1;
-        }
-      }
-    }
-    &--plus {
-      background-color: $green-500;
-      &::before {
-        @include p_center-all;
-        width: $size_icon;
-        height: 2px;
-        content: "";
-        border-radius: 2px;
-        background-color: $white;
-      }
-      &::after {
-        @include p_center-all;
-        width: $size_icon;
-        height: 2px;
-        content: "";
-        transform: translate(-50%, -50%) rotate(90deg);
-        border-radius: 2px;
-        background-color: $white;
-      }
-      &:hover:not(:active):not(:disabled) {
-        background-color: $green-400;
-      }
-      &:active:not(:disabled) {
-        background-color: $green-600;
-      }
-      &:focus:not(:disabled) {
-        box-shadow: $shadow-regular;
-      }
-      &:disabled {
-        cursor: default;
-        opacity: 0.3;
-      }
-    }
-    &--orange {
-      background-color: $orange-200;
-      &:hover:not(:active):not(:disabled) {
-        background-color: $orange-100;
-      }
-      &:active:not(:disabled) {
-        background-color: $orange-300;
-      }
-    }
-  }
-  .counter__input {
-    @include r-s14-h16;
-    box-sizing: border-box;
-    width: 22px;
-    margin: 0;
-    padding: 0 3px;
-    text-align: center;
-    color: $black;
-    border: none;
-    border-radius: 10px;
-    outline: none;
-    background-color: transparent;
-    &:focus {
-      box-shadow: inset $shadow-regular;
-    }
-  }
-  .dough__input {
-    position: relative;
-    margin-right: 8%;
-    margin-bottom: 20px;
-    padding-left: 50px;
-    cursor: pointer;
-    img {
-      @include p_center-v;
-      width: 36px;
-      height: 36px;
-      transition: 0.3s;
-      border-radius: 50%;
-    }
-    b {
-      @include r-s16-h19;
-    }
-    span {
-      @include l-s11-h13;
-      display: block;
-    }
-    &:hover {
-      img {
-        box-shadow: $shadow-regular;
-      }
-    }
-    input {
-      &:checked + img {
-        box-shadow: $shadow-large;
-      }
-    }
-  }
-  .diameter__input {
-    margin-right: 8.7%;
-    margin-bottom: 20px;
-    padding-top: 7px;
-    padding-bottom: 6px;
-    cursor: pointer;
-    span {
-      @include r-s16-h19;
-      position: relative;
-      padding-left: 46px;
-      &::before {
-        @include p_center_v;
-        width: 36px;
-        height: 36px;
-        content: "";
-        transition: 0.3s;
-        border-radius: 50%;
-        background-color: $green-100;
-        background-image: url("@/assets/img/diameter.svg");
-        background-repeat: no-repeat;
-        background-position: center;
-      }
-    }
-    &:nth-child(3n) {
-      margin-right: 0;
-    }
-    &--small {
-      span::before {
-        background-size: 18px;
-      }
-    }
-    &--normal {
-      span::before {
-        background-size: 29px;
-      }
-    }
-    &--big {
-      span::before {
-        background-size: 100%;
-      }
-    }
-    &:hover {
-      span::before {
-        box-shadow: $shadow-regular;
-      }
-    }
-    input {
-      &:checked + span::before {
-        box-shadow: $shadow-large;
-      }
-    }
-  }
-  .filling {
-    @include r-s14-h16;
-    position: relative;
-    display: block;
-    padding-left: 36px;
-    img {
-      @include p_center-v;
-      display: block;
-      width: 32px;
-      height: 32px;
-      box-sizing: border-box;
-      padding: 4px;
-      border-radius: 50%;
-    }
-  }
-  .button {
-    $bl: &;
-    @include b-s18-h21;
-    font-family: inherit;
-    display: block;
-    box-sizing: border-box;
-    margin: 0;
-    padding: 0;
-    cursor: pointer;
-    transition: 0.3s;
-    text-align: center;
-    color: $white;
-    border: none;
-    border-radius: 8px;
-    outline: none;
-    box-shadow: $shadow-medium;
-    background-color: $green-500;
-    &:hover:not(:active):not(:disabled) {
-      background-color: $green-400;
-    }
-    &:active:not(:disabled) {
-      background-color: $green-600;
-    }
-    &:focus:not(:disabled) {
-      opacity: 0.5;
-    }
-    &:disabled {
-      background-color: $green-300;
-      color: rgba($white, 0.2);
-      cursor: default;
-    }
-    &--border {
-      background-color: transparent;
-      border: 1px solid $green-500;
-      color: $black;
-      box-shadow: none;
-      &:hover:not(:active):not(:disabled) {
-        color: $green-500;
-        border-color: $green-500;
-        background-color: transparent;
-      }
-      &:active:not(:disabled) {
-        color: $green-600;
-        border-color: $green-600;
-        background-color: transparent;
-      }
-      &:disabled {
-        opacity: 0.5;
-      }
-    }
-    &--transparent {
-      @include b-s14-h16;
-      background-color: transparent;
-      box-shadow: none;
-      color: $black;
-      &:hover:not(:active):not(:disabled) {
-        color: $red-800;
-        background-color: transparent;
-      }
-      &:active:not(:disabled) {
-        color: $red-900;
-        background-color: transparent;
-      }
-      &:disabled {
-        opacity: 0.25;
-      }
-    }
-    &--arrow {
-      &::before {
-        content: "";
-        background-image: url("@/assets/img/button-arrow.svg");
-        background-position: center;
-        background-repeat: no-repeat;
-        margin-right: 16px;
-        width: 18px;
-        height: 18px;
-        display: inline-block;
-        vertical-align: middle;
-        transform: translateY(-1px);
-      }
-    }
-    &--white {
-      background-color: $white;
-      color: $green-500;
-    }
-  }
-  .pizza {
-    position: relative;
-    display: block;
-    box-sizing: border-box;
-    width: 100%;
-    background-repeat: no-repeat;
-    background-position: center;
-    background-size: 100%;
-    &--foundation--big-creamy {
-      background-image: url("@/assets/img/foundation/big-creamy.svg");
-    }
-    &--foundation--big-tomato {
-      background-image: url("@/assets/img/foundation/big-tomato.svg");
-    }
-    &--foundation--small-creamy {
-      background-image: url("@/assets/img/foundation/small-creamy.svg");
-    }
-    &--foundation--small-tomato {
-      background-image: url("@/assets/img/foundation/small-tomato.svg");
-    }
-  }
-  .pizza__wrapper {
-    width: 100%;
-    padding-bottom: 100%;
-  }
-  .pizza__filling {
-    $bl: &;
-    position: absolute;
-    top: 0;
-    left: 0;
-    display: block;
-    width: 100%;
-    height: 100%;
-    background-repeat: no-repeat;
-    background-position: center;
-    background-size: 100%;
-    &::before,
-    &::after {
-      display: none;
-      position: absolute;
-      top: 0;
-      left: 0;
-      width: 100%;
-      height: 100%;
-      content: "";
-      background-image: inherit;
-    }
-    &--second {
-      &::before {
-        display: block;
-        transform: rotate(45deg);
-      }
-    }
-    &--third {
-      &::before {
-        display: block;
-        transform: rotate(45deg);
-      }
-      &::after {
-        display: block;
-        transform: rotate(-45deg);
-      }
-    }
-    &--ananas,
-    &--ananas.pizza__filling--second::before,
-    &--ananas.pizza__filling--third::after {
-      background-image: url("@/assets/img/filling-big/ananas.svg");
-    }
-    &--bacon,
-    &--bacon.pizza__filling--second::before,
-    &--bacon.pizza__filling--third::after {
-      background-image: url("@/assets/img/filling-big/bacon.svg");
-    }
-    &--blue_cheese,
-    &--blue.pizza__filling--second::before,
-    &--blue.pizza__filling--third::after {
-      background-image: url("@/assets/img/filling-big/blue_cheese.svg");
-    }
-    &--cheddar,
-    &--cheddar.pizza__filling--second::before,
-    &--cheddar.pizza__filling--third::after {
-      background-image: url("@/assets/img/filling-big/cheddar.svg");
-    }
-    &--chile,
-    &--chile.pizza__filling--second::before,
-    &--chile.pizza__filling--third::after {
-      background-image: url("@/assets/img/filling-big/chile.svg");
-    }
-    &--ham,
-    &--ham.pizza__filling--second::before,
-    &--ham.pizza__filling--third::after {
-      background-image: url("@/assets/img/filling-big/ham.svg");
-    }
-    &--jalapeno,
-    &--jalapeno.pizza__filling--second::before,
-    &--jalapeno.pizza__filling--third::after {
-      background-image: url("@/assets/img/filling-big/jalapeno.svg");
-    }
-    &--mozzarella,
-    &--mozzarella.pizza__filling--second::before,
-    &--mozzarella.pizza__filling--third::after {
-      background-image: url("@/assets/img/filling-big/mozzarella.svg");
-    }
-    &--mushrooms,
-    &--mushrooms.pizza__filling--second::before,
-    &--mushrooms.pizza__filling--third::after {
-      background-image: url("@/assets/img/filling-big/mushrooms.svg");
-    }
-    &--olives,
-    &--olives.pizza__filling--second::before,
-    &--olives.pizza__filling--third::after {
-      background-image: url("@/assets/img/filling-big/olives.svg");
-    }
-    &--onion,
-    &--onion.pizza__filling--second::before,
-    &--onion.pizza__filling--third::after {
-      background-image: url("@/assets/img/filling-big/onion.svg");
-    }
-    &--parmesan,
-    &--parmesan.pizza__filling--second::before,
-    &--parmesan.pizza__filling--third::after {
-      background-image: url("@/assets/img/filling-big/parmesan.svg");
-    }
-    &--salami,
-    &---salami.pizza__filling--second::before,
-    &---salami.pizza__filling--third::after {
-      background-image: url("@/assets/img/filling-big/salami.svg");
-    }
-    &--salmon,
-    &--salmon.pizza__filling--second::before,
-    &--salmon.pizza__filling--third::after {
-      background-image: url("@/assets/img/filling-big/salmon.svg");
-    }
-    &--tomatoes,
-    &--tomatoes.pizza__filling--second::before,
-    &--tomatoes.pizza__filling--third::after {
-      background-image: url("@/assets/img/filling-big/tomatoes.svg");
-    }
-  }
-  .input {
-    display: block;
-    span {
-      @include r-s14-h16;
-      display: block;
-      margin-bottom: 4px;
-    }
-    input {
-      @include r-s16-h19;
-      display: block;
-      box-sizing: border-box;
-      width: 100%;
-      margin: 0;
-      padding: 8px 16px;
-      transition: 0.3s;
-      color: $black;
-      border: 1px solid $purple-400;
-      border-radius: 8px;
-      outline: none;
-      background-color: $white;
-      font-family: inherit;
-      &:focus {
-        border-color: $green-500;
-      }
-    }
-    &:hover {
-      input {
-        border-color: $black;
-      }
-    }
-    &--big-label {
-      display: flex;
-      align-items: center;
-      span {
-        @include b-s16-h19;
-        margin-right: 16px;
-        white-space: nowrap;
-      }
-    }
-  }
-  
-  
-  
 
 </style>
